o_classifier/requirements.txt
numpy
Flask-API
pickle-mixin